#155 Validate `output` parameter values
